+++ encoders/dem_rcf.py
@@ -36,22 +36,8 @@ def preprocess_dem(dem_data):
 
 
 def get_embedding(dem_data, model):
-    """
-    Generate embeddings for DEM data using torchgeo's RandomFeatureConvolution model.
-
-    Args:
-        dem_data (numpy.ndarray): 2D array representing the DEM data.
-
-    Returns:
-        torch.Tensor: Embeddings generated by the RCF model.
-    """
     # Convert the DEM data to a PyTorch tensor
     dem_tensor = dem_data.unsqueeze(0)
-    # (
-    #     torch.tensor(dem_data, dtype=torch.float32).unsqueeze(0)
-    # )  # Shape: (1, 1, H, W)
-
-    # print(dem_tensor.shape)
 
     with torch.no_grad():
         embeddings = model(dem_tensor)
@@ -60,12 +46,7 @@ def get_embedding(dem_data, model):
     return embeddings.flatten()
 
 
-if __name__ == "__main__":
-    # directory = Config.DATA_DIR_LBL_DEM
-    directory = Config.DATA_DIR_UNLBL_DEM
-    embeddings_dir = "data/unlabeled/embeddings/dem_v2"
-    print(directory)
-
+def main(directory, embeddings_dir):
     dem_dataset = DEMDataset()
     # Initialize the RCF model
     model = RCF(
@@ -92,7 +73,16 @@ if __name__ == "__main__":
 
             embedding = get_embedding(dem_data, model)
             # print(embedding)
-            # print(encode_file(lat, lon, "dem", embeddings_dir, "npy"))
-            np.save(
-                encode_file(lat, lon, "dem", embeddings_dir, "npy"), embedding.numpy()
-            )
+            embeddings_file_path = encode_file(lat, lon, "dem", embeddings_dir, "npy")
+            # print(embeddings_file_path)
+            np.save(embeddings_file_path, embedding.numpy())
+
+
+if __name__ == "__main__":
+    directory = Config.DATA_DIR_LBL_DEM
+    embeddings_dir = "data/labeled/embeddings/dem_v2"
+    main(directory, embeddings_dir)
+
+    directory = Config.DATA_DIR_UNLBL_DEM
+    embeddings_dir = "data/unlabeled/embeddings/dem_v2"
+    main(directory, embeddings_dir)
